PB-1515 : Move AssetAdminMixin to admin.py & address various review comments
